feat: split ApiGatewayHandler
